+++ src/main/java/ladder/present/Renderer.java
@@ -1,5 +1,12 @@
-package ladder.domain;
+package ladder.present;
 
+
+import ladder.domain.Column;
+import ladder.domain.Ladder;
+import ladder.domain.Line;
+import ladder.domain.Row;
+import ladder.domain.Scene;
+import ladder.domain.Users;
 
 import java.util.List;
 import java.util.stream.Collectors;
